Updated implementation of “overwrite_files” mode
